Add warning for configured app in dropins
