+++ dev/com.ibm.ws.app.manager_fat/fat/src/componenttest/application/manager/test/FATSuite.java
@@ -18,7 +18,8 @@ import org.junit.runners.Suite.SuiteClasses;
 @SuiteClasses({
                 FATTest.class,
                 AutoExtractTest.class,
-                AppOrderTests.class
+                AppOrderTests.class,
+                DropinsTests.class
 })
 public class FATSuite {
