+++ dev/com.ibm.ws.app.manager_fat/publish/files/dataSourceApp/server.xml
@@ -5,8 +5,8 @@
     <applicationMonitor dropins="dropins(fish)"/>
 
     <featureManager>
-       <feature>ejbLite-3.1</feature>
-       <feature>jdbc-4.0</feature>
+       <feature>ejbLite-3.2</feature>
+       <feature>jdbc-4.1</feature>
     </featureManager>
 
     <!-- Derby-only JDBC driver -->
