+++ dev/com.ibm.ws.app.manager_fat/publish/servers/dropinsTestServer/bootstrap.properties
@@ -1,0 +1,3 @@
+bootstrap.include=../testports.properties
+com.ibm.ws.logging.trace.specification=*=event=enabled
+com.ibm.ws.logging.max.file.size=0
