+++ dev/com.ibm.ws.app.manager_fat/publish/servers/dropinsTestServer/server.xml
@@ -1,0 +1,11 @@
+<server description="default server">
+
+    <include location="../fatTestPorts.xml"/>
+
+	<applicationMonitor dropins="dropins(fish)"/>
+
+    <featureManager>
+	   <feature>servlet-3.1</feature>
+    </featureManager>
+    
+</server>
